# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/app/address/add": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "新增地址",
                "parameters": [
                    {
                        "type": "string",
                        "description": "收货人名称",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户open_id",
                        "name": "open_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "行政区域表的省",
                        "name": "province",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "行政区域表的市",
                        "name": "city",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "行政区域表的区县",
                        "name": "county",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "详细收货地址",
                        "name": "detail",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "手机号码",
                        "name": "tel",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "是否默认地址",
                        "name": "is_default",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/app/address/delete": {
            "delete": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "删除地址",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "地址id",
                        "name": "address_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "open_id",
                        "name": "open_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/app/address/detail": {
            "get": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取地址详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "open_id",
                        "name": "open_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "地址id",
                        "name": "address_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/app.Address"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/app/address/list": {
            "get": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取地址列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "open_id",
                        "name": "open_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/app.Address"
                            }
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/app/address/update": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "更改地址",
                "parameters": [
                    {
                        "type": "string",
                        "description": "收货人名称",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户open_id",
                        "name": "open_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "行政区域表的省",
                        "name": "province",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "行政区域表的市",
                        "name": "city",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "行政区域表的区县",
                        "name": "county",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "详细收货地址",
                        "name": "detail",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "手机号码",
                        "name": "tel",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "是否默认地址",
                        "name": "is_default",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app.Address": {
            "type": "object",
            "properties": {
                "city": {
                    "description": "行政区域表的市ID",
                    "type": "string"
                },
                "county": {
                    "description": "行政区域表的区县ID",
                    "type": "string"
                },
                "created_time": {
                    "description": "创建时间",
                    "type": "string"
                },
                "deleted": {
                    "description": "逻辑删除",
                    "type": "integer"
                },
                "deleted_time": {
                    "description": "删除时间",
                    "type": "string"
                },
                "detail": {
                    "description": "详细收货地址",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_default": {
                    "description": "是否默认地址",
                    "type": "integer"
                },
                "name": {
                    "description": "收货人名称",
                    "type": "string"
                },
                "open_id": {
                    "description": "用户表的用户ID",
                    "type": "string"
                },
                "province": {
                    "description": "行政区域表的省ID",
                    "type": "string"
                },
                "tel": {
                    "description": "手机号码",
                    "type": "string"
                },
                "updated_time": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}